Updating reservation

+++ app/_lib/actions.js
@@ -4,6 +4,7 @@ import { supabase } from './supabase';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 import { auth } from './auth';
+import { getBookings } from './data-service';
 
 export async function signInAction() {
   // this won't work for popup
@@ -46,6 +47,13 @@ export async function deleteReservation(bookingId) {
   const session = await auth();
   if (!session) throw new Error('You must be logged in to delete reservations');
 
+  const guestBookings = await getBookings(session.user.guestId);
+
+  const guestBookingsId = guestBookings.map((booking) => booking.id);
+  if (!guestBookingsId.includes(bookingId)) {
+    throw new Error('You are not authorized to delete this reservation');
+  }
+
   const { error } = await supabase
     .from('bookings')
     .delete()
@@ -54,4 +62,33 @@ export async function deleteReservation(bookingId) {
   if (error) throw new Error('Reservation could not be deleted');
 
   revalidatePath('/account/reservations');
+}
+
+export async function updateBooking(formData) {
+  console.log(formData);
+  const session = await auth();
+  if (!session) throw new Error('You must be logged in to update your booking');
+
+  const guestBookings = await getBookings(session.user.guestId);
+  const bookingId = Number(formData.get('bookingId'));
+
+  const guestBookingsId = guestBookings.map((booking) => booking.id);
+  if (!guestBookingsId.includes(bookingId))
+    throw new Error('You are not authorized to update this reservation');
+
+  const updateFields = {
+    numGuests: Number(formData.get('numGuests')),
+    observations: formData.get('observations').slice(0, 1000),
+  };
+
+  const { error } = await supabase
+    .from('bookings')
+    .update(updateFields)
+    .eq('id', bookingId)
+    .select()
+    .single();
+
+  if (error) throw new Error('Booking could not be updated');
+
+  redirect('/account/reservations');
 }
